+++ src/CodecTactics.MonoGame/Program.cs
@@ -1,0 +1,2 @@
+using var game = new CodecTactics.MonoGame.Game1();
+game.Run();